+++ concordium-android-wallet-example/app/src/main/java/com/concordium/example/wallet/Storage.kt
@@ -3,8 +3,6 @@ package com.concordium.example.wallet
 import android.content.Context
 import android.content.SharedPreferences
 import androidx.activity.ComponentActivity
-import cash.z.ecc.android.bip39.Mnemonics
-import cash.z.ecc.android.bip39.toSeed
 import com.concordium.sdk.crypto.wallet.ConcordiumHdWallet
 
 /***
@@ -23,11 +21,9 @@ class Storage(context: Context) {
     val accountAddress = StorageAccessor(sharedPreferences, "account_address")
     val identity = StorageAccessor(sharedPreferences, "identity")
 
-    @OptIn(ExperimentalStdlibApi::class)
     fun getWallet(): ConcordiumHdWallet {
         val seedPhrase = this.seedPhrase.get()
-        val seedAsHex = Mnemonics.MnemonicCode(seedPhrase!!.toCharArray()).toSeed().toHexString()
-        return ConcordiumHdWallet.fromHex(seedAsHex, Constants.NETWORK)
+        return ConcordiumHdWallet.fromSeedPhrase(seedPhrase, Constants.NETWORK)
     }
 }